requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
passlib[bcrypt]
PyJWT
pydantic
